+++ crates/eon-manifest/src/lib.rs
@@ -464,7 +464,7 @@ mod tests {
         assert_eq!(
             report,
             "eon-alpha x86_64-linux\n\
-orbit 0.1.0 69c402737799f03e615473956954a043647a4713 x86_64-linux\n\
+orbit 0.1.0 0ca0cc93b83793d08eadcca7eac6e947a80ce25d x86_64-linux\n\
 venus 0.1.0 3612a929402de00d07d582b6547d99f66afe5cad x86_64-linux\n\
 nushell 0.113.1 7b7df4aa68e957cf38b9d8157c35fa7523f44a6d x86_64-linux\n\
 bash 5.3p9 b8c60bc9ca365f8261fa97900b6fa939f6ebc303 x86_64-linux\n\
@@ -482,7 +482,7 @@ ratconfig 6.0.0 e6ec2ebfe84b2358186410680cbcaf0564eb59a2 x86_64-linux"
         assert!(!report.contains("/nix/store"));
         assert_eq!(
             component_revision(CANONICAL, "orbit").unwrap(),
-            "69c402737799f03e615473956954a043647a4713"
+            "0ca0cc93b83793d08eadcca7eac6e947a80ce25d"
         );
         assert!(component_revision(CANONICAL, "missing").is_err());
     }